refactor(*): move `GameActionClient` to `clients`
Ref: #575
