+++ src-frontend-app-web/src/clients/index.ts
@@ -1,5 +1,6 @@
 export * from "./character-creation.client";
 export * from "./configuration.client";
+export * from "./game-action.client";
 export * from "./hall-of-fame.client";
 export * from "./localization.client";
 export * from "./login.client";
